+++ src/extension.ts
@@ -2,93 +2,129 @@
 
 import * as vscode from 'vscode';
 import { GNURadioController } from './controller';
+import { exec, execOnFile } from './shellTask';
 
 export function activate(context: vscode.ExtensionContext) {
-    const ctl = new GNURadioController(context);
+    const extId: string = context.extension.packageJSON.name;
 
-    if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length > 0) {
-        ctl.setCwd(vscode.workspace.workspaceFolders[0].uri.fsPath);
-    }
+    const getConfig = <T>(key: string) => vscode.workspace.getConfiguration(extId).get<T>(key);
+
+    const getWorkspaceDir = () => vscode.workspace.workspaceFolders?.length === 1
+        ? vscode.workspace.workspaceFolders[0].uri.fsPath : undefined;
+
+    /** 
+     * Open GNURadio Companion application.
+     * 
+     * This command runs `gnuradio-companion` in the shell.
+     */
+    const openGnuradioCompanion = vscode.commands.registerCommand(
+        `${extId}.openGnuradioCompanion`,
+        () => {
+            const cmd = getConfig<string>('companion.cmd') ?? 'gnuradio-companion';
+            exec(`"${cmd}"`, getWorkspaceDir());
+        });
+
+    /** 
+     * Edit the file in GNURadio Companion application.
+     * 
+     * This command runs `gnuradio-companion %f` in the shell, opening the selected file `%f`.
+     */
+    const editInGnuradioCompanion = vscode.commands.registerCommand(
+        `${extId}.editInGnuradioCompanion`,
+        (fileUri?: vscode.Uri) => {
+            const cmd = getConfig<string>('companion.cmd') ?? 'gnuradio-companion';
+            execOnFile(`"${cmd}"`, fileUri, '.grc');
+        });
+
+    /**
+     * Compile the GRC flowgraph file.
+     * 
+     * This command runs `grcc %f` in the shell, producing a Python executable in the same folder as the selected file `%f`.
+     */
+    const compileFlowgraph = vscode.commands.registerCommand(
+        `${extId}.compileFlowgraph`,
+        (fileUri?: vscode.Uri) => {
+            const cmd = getConfig<string>('compiler.cmd') ?? 'grcc';
+            execOnFile(`"${cmd}"`, fileUri, '.grc');
+        });
+
+    /**
+     * Run the GRC flowgraph file.
+     * 
+     * This command runs `grcc -r %f` in the shell, producing a Python executable in the same folder as the selected file `%f` and running it.
+     */
+    const runFlowgraph = vscode.commands.registerCommand(
+        `${extId}.runFlowgraph`,
+        (fileUri?: vscode.Uri) => {
+            const cmd = getConfig<string>('compiler.cmd') ?? 'grcc';
+            execOnFile(`"${cmd}" -r`, fileUri, '.grc');
+        });
+
+    const ctl = new GNURadioController(context);
+    ctl.setCwd(getWorkspaceDir());
 
     context.subscriptions.push(
-        vscode.workspace.onDidChangeWorkspaceFolders((e) => {
-            if (e.added.length) {
-                ctl.setCwd(e.added[0].uri.fsPath);
-            } else if (e.removed.length) {
-                ctl.setCwd();
-            }
+        openGnuradioCompanion,
+        editInGnuradioCompanion,
+        compileFlowgraph,
+        runFlowgraph,
+        ctl,
+        vscode.workspace.onDidChangeWorkspaceFolders(_ => {
+            ctl.setCwd(getWorkspaceDir());
         }),
-        vscode.workspace.onDidChangeConfiguration((e) => {
+        vscode.workspace.onDidChangeConfiguration(e => {
             if (e.affectsConfiguration('gnuradio-integration.modtool.checkXml') &&
-                vscode.workspace.getConfiguration().get('gnuradio-integration.modtool.checkXml') === true) {
+                getConfig<boolean>('modtool.checkXml') === true) {
                 ctl.checkXml();
             }
         }),
         vscode.tasks.onDidEndTaskProcess(e => {
-            if (e.execution.task.source === 'grc') {
-                if (e.exitCode && e.exitCode !== 0) {
-                    vscode.window.showErrorMessage(
-                        `Compilation process finished with error code ${e.exitCode}; check the terminal output for details`);
-                } else if (e.execution.task.detail?.startsWith('grcc')) {
-                    // TODO: C++ flowgraph?
-                    // TODO: read task parameters to find the compiled file
-                    vscode.window.showInformationMessage('Flowgraph compilation was successfull');
-                }
+            if (e.execution.task.source !== 'gnuradio') {
+                return;
+            }
+            if (e.exitCode && e.exitCode !== 0) {
+                vscode.window.showErrorMessage(
+                    `Task finished with error code ${e.exitCode}; check the terminal output for details`);
+            } else if (e.execution.task.detail?.startsWith('grcc')) {
+                // TODO: C++ flowgraph?
+                // TODO: read task parameters to find the compiled file
+                vscode.window.showInformationMessage('Flowgraph compilation was successfull');
             }
         }),
-    );
-
-    context.subscriptions.push(
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.openGnuradioCompanion.name}`,
-            ctl.openGnuradioCompanion,
-            ctl),
-        vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.editInGnuradioCompanion.name}`,
-            ctl.editInGnuradioCompanion,
-            ctl),
-        vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.runFlowgraph.name}`,
-            ctl.runFlowgraph,
-            ctl),
-        vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.compileFlowgraph.name}`,
-            ctl.compileFlowgraph,
-            ctl),
-        vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.createModule.name}`,
+            `${extId}.${ctl.createModule.name}`,
             ctl.createModule,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.getModuleInfo.name}`,
+            `${extId}.${ctl.getModuleInfo.name}`,
             ctl.getModuleInfo,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.createBlock.name}`,
+            `${extId}.${ctl.createBlock.name}`,
             ctl.createBlock,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.createPythonBindings.name}`,
+            `${extId}.${ctl.createPythonBindings.name}`,
             ctl.createPythonBindings,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.renameBlock.name}`,
+            `${extId}.${ctl.renameBlock.name}`,
             ctl.renameBlock,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.disableBlock.name}`,
+            `${extId}.${ctl.disableBlock.name}`,
             ctl.disableBlock,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.removeBlock.name}`,
+            `${extId}.${ctl.removeBlock.name}`,
             ctl.removeBlock,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.convertXmlToYaml.name}`,
+            `${extId}.${ctl.convertXmlToYaml.name}`,
             ctl.convertXmlToYaml,
             ctl),
         vscode.commands.registerCommand(
-            `${ctl.extId}.${ctl.makeYamlFromImpl.name}`,
+            `${extId}.${ctl.makeYamlFromImpl.name}`,
             ctl.makeYamlFromImpl,
             ctl),
     );
@@ -102,7 +138,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     const registerTreeItemAlias = (alias: string, command: string) =>
         vscode.commands.registerCommand(
-            `${ctl.extId}.${alias}`,
+            `${extId}.${alias}`,
             (item?: vscode.TreeItem) => {
                 if (!item) {
                     if (!moduleView.selection.length) {
@@ -120,7 +156,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     context.subscriptions.push(
         vscode.commands.registerCommand(
-            `${ctl.extId}.refreshView`,
+            `${extId}.refreshView`,
             ctl.refresh,
             ctl),
         registerTreeItemAlias('fileOpenBeside', 'explorer.openToSide'),